.history/requirements_20250520192033.txt
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3.post1
pandas_ta==0.3.14b0
pytest==7.4.3
matplotlib==3.7.2
